--- Backend/Haskell/src/Imports.hs
@@ -1,8 +1,0 @@
-{-# LANGUAGE ConstraintKinds #-}
-module Imports
-  ( Has 
-  ) where
-
-import Control.Effect.Carrier (Carrier, Member)
-
-type Has eff sig m = (Carrier sig m, Member eff sig)
